requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langsmith==0.1.69
lark==1.1.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
onnxruntime==1.18.0
openai==1.30.1
opencv-python==4.10.0.82
orjson==3.10.3
packaging==23.2
pathlib==1.0.1
pillow==10.3.0
pinecone-client==3.2.2
pinecone-notebooks==0.1.1
pinecone-text==0.9.0
protobuf==5.27.0
psycopg2-binary==2.9.9
pyclipper==1.3.0.post5
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
pypdf==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidocr-onnxruntime==1.3.22
regex==2024.5.15
requests==2.32.3
rich==13.7.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12.1
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
weasyprint==60.2
Werkzeug==3.0.3
wget==3.2
yarl==1.9.4
EbookLib==0.18
moviepy>=1.0.3
openai-whisper>=20230314
PyMuPDF==1.25.5
pinecone==3.2.2
elevenlabs==1.57.0
PyJWT==2.8.0
bcrypt==4.1.2
boto3==1.34.69